requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiohttp==3.9.1
    # via cohere
aiosignal==1.3.1
    # via aiohttp
attrs==23.2.0
    # via aiohttp
autopep8==2.0.4
    # via -r requirements.in
backoff==2.2.1
    # via cohere
beautifulsoup4==4.12.2
    # via -r requirements.in
build==1.0.3
    # via pip-tools
certifi==2023.11.17
    # via requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via pip-tools
cohere==4.40
    # via -r requirements.in
colorama==0.4.6
    # via
    #   build
    #   click
    #   tqdm
distlib==0.3.8
    # via virtualenv
fastavro==1.9.2
    # via cohere
filelock==3.13.1
    # via virtualenv
flake8==6.1.0
    # via -r requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
identify==2.5.33
    # via pre-commit
idna==3.6
    # via
    #   requests
    #   yarl
importlib-metadata==6.11.0
    # via cohere
lxml==4.9.3
    # via -r requirements.in
mccabe==0.7.0
    # via flake8
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
nodeenv==1.8.0
    # via pre-commit
packaging==23.2
    # via build
pip-tools==7.3.0
    # via -r requirements.in
platformdirs==4.1.0
    # via virtualenv
pre-commit==3.6.0
    # via -r requirements.in
pycodestyle==2.11.1
    # via
    #   autopep8
    #   flake8
pyflakes==3.1.0
    # via flake8
pyproject-hooks==1.0.0
    # via build
python-dotenv==1.0.0
    # via -r requirements.in
pyyaml==6.0.1
    # via pre-commit
requests==2.31.0
    # via
    #   -r requirements.in
    #   cohere
soupsieve==2.5
    # via beautifulsoup4
tqdm==4.66.1
    # via -r requirements.in
urllib3==2.1.0
    # via
    #   cohere
    #   requests
virtualenv==20.25.0
    # via
    #   -r requirements.in
    #   pre-commit
wheel==0.42.0
    # via pip-tools
yarl==1.9.4
    # via aiohttp
zipp==3.17.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
